separated routes and added deploy bash script
