+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "body-parser": "^1.19.0",
     "dotenv": "^8.2.0",
+    "ejs": "^3.1.5",
     "express": "^4.17.1",
     "firebase": "^8.2.1",
     "firebase-admin": "^9.4.2",
