+++ public/index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <title>Comet Data Service Admin Console</title>
     <meta name="google-signin-client_id"
-        content="***REMOVED***-5csqgrjl6raqo5q52ohrdp6alr9ef8ol.apps.googleusercontent.com">
+        content="<%= OAUTH_CLIENT_ID %>">
     <meta name="google-signin-cookiepolicy" content="single_host_origin">
     <meta name="google-signin-scope" content="profile email">
     <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -151,17 +151,17 @@
     <script>
         // Your web app's Firebase configuration
         var firebaseConfig = {
-            apiKey: "***REMOVED***",
-            authDomain: "cometplanning.firebaseapp.com",
+            apiKey: "<%= FIREBASE_API_KEY %>", 
+            authDomain: "cometplanning.firebaseapp.com", 
             databaseURL: "https://cometplanning.firebaseio.com",
             projectId: "cometplanning",
             storageBucket: "cometplanning.appspot.com",
-            messagingSenderId: "***REMOVED***",
-            appId: "1:***REMOVED***:web:1c12ccb40998f2f3492342"
+            messagingSenderId: "<%= FIREBASE_MESSAGING_ID %>",
+            appId: "<%= FIREBASE_APP_ID %>"
         };
         // Initialize Firebase
         firebase.initializeApp(firebaseConfig);
-    </script>
+    </script> 
     <script src="https://apis.google.com/js/platform.js?onload=onLoad" async defer></script>
     <script src="auth.js"></script>
     <script src="index.js"></script>
